split styles file
