+++ jsapp/js/components/processing/sidebar/sidebarSubmissionData.module.scss
@@ -1,0 +1,14 @@
+@use '~kobo-common/src/styles/colors';
+@use 'scss/sizes';
+
+.dataList {
+  flex: 1;
+  padding: sizes.$x12;
+  border: sizes.$x1 solid colors.$kobo-gray-85;
+  border-radius: sizes.$x6;
+}
+
+.dataListBody {
+  overflow-y: auto;
+  height: 100%;
+}
