+++ jsapp/js/components/processing/sidebar/sidebarSubmissionData.tsx
@@ -1,9 +1,9 @@
 import React, {useState} from 'react';
 import SubmissionDataList from 'js/components/submissions/submissionDataList';
 import singleProcessingStore from 'js/components/processing/singleProcessingStore';
-import {AssetContent} from 'jsapp/js/dataInterface';
+import type {AssetContent} from 'js/dataInterface';
 import {META_QUESTION_TYPES, ADDITIONAL_SUBMISSION_PROPS} from 'js/constants';
-import styles from './sidebarSubmissionDetails.module.scss';
+import styles from './sidebarSubmissionData.module.scss';
 
 interface SidebarSubmissionDataProps {
   asset: AssetContent | undefined;
