+++ jsapp/js/components/processing/sidebar/sidebarSubmissionMedia.module.scss
@@ -1,6 +1,5 @@
 @use '~kobo-common/src/styles/colors';
 @use 'scss/sizes';
-@use '../singleProcessingVariables';
 
 .mediaWrapper {
   &.mediaWrapperAudio {
@@ -14,18 +13,6 @@
     border-radius: sizes.$x6;
     overflow: hidden;
   }
-}
-
-.dataList {
-  flex: 1;
-  padding: sizes.$x12;
-  border: sizes.$x1 solid colors.$kobo-gray-85;
-  border-radius: sizes.$x6;
-}
-
-.dataListBody {
-  overflow-y: auto;
-  height: 100%;
 }
 
 .videoPreview {
